[seq/event_time] Class to represent time of an event in basic sequence
With partial sorting and optimizations.

The time is represented as a sum of multiple terms
with some assumptions of the signs of each terms.
This is enough to identify most of the known orders between the times (by comparing terms).
However, this will not be enough to handle waiting or manually positioned pulses.
That will be handled by explicitly specifying the ordering relation using frontend knowledge.