[seq/pulse] Pulse class
Representing outputs or measures.

This does **NOT** manage the (Event)Time representing the start of the pulse.
There will likely be multiple pulses with exactly the same time
and managing the times outside of the pulses allows those to be identified by pointer identity
as well as optimizing each time only once.

The Pulse class also handles the tracking of the pulse end value.