[seq/basic_seq] Sequence without branches
The name basic sequence is borrowed from basic block in compilers.
This hosts most of the information about a sequence including pulses, measures,
lengths, branch conditions and targets and miscellaneous things like
assumptions and automatic assignment of globals.
Like basic blocks, this is also the abstraction that host most of the information
that we can easily reason about and do optimization and transformation with.
These includes,

* Time order discovery
* Endtime/pulse ordering
* Measurement/pulse start value forwarding
* Dead branch elimination
* Unused time elimination

This is also the class that performs checks on the sequence to make sure it's valid
especially regarding the use of measurement values: they can only be used in the same
basic sequence on a pulse after when they are measured.