[seq/seq] Sequence with control flow
Built on top of the basic sequence, this manages the basic sequences
that are within a sequence connected with control flow.
Since most of the sequence is actually handled by basic sequence,
this only deal with global states and the interaction between basic sequences including,

* Channel management
* Default values
* Global variable management
* Dead basic sequence elimination
* Constant value forwarding between sequences

This is also the class that drives the whole optimization pipeline.
The user will only use this class (instead of the transformation/optimization functions
in the time, pulse or basic sequence classes) to prepare and optimize the whole sequence.