[seq/builder] Construct sequence from serialization
Unlike the previous implementation for the FPGA backend,
we do not directly use the serialization for `IR::Function`s.
Rather, each values and their dependencies are passed in directly
and we'll construct the proper functions from there.
This give us more informations about the intermediate values to help with optimizations.