[seq/basic_seq] Create local variables for the start values of the basic sequence and each pulse
This allows us to optimize the values in the sequence slightly better.
This also thread the value through the sequence such that the implicit value dependency
from the pulse relative timing can be represented in the variable graph directly.
The latter does not allow much values to be precomputed before runtime
but should help more evaluation order to be statically determined.