[seq/zynq] Sequence version 2 with shorter minimum time
Update cmdline tools

The new FPGA code use pipelining to shorten the minimum time of a pulse so that a TTL pulse
can take as short as only one cycle now.
Since this minimum time is encoded in the bytecode format directly,
we need a new bytecode version to take advantage of it.