[seq/zynq/bc_gen] Bytecode generator for the Zynq backend of the new sequence implementation.
This is slightly simplified from the legacy implementation by dividing the work into more passes.
For example, the rate limitation on the pulses are now done before we start scheduling the pulses
and so is the conversion from clock period to clock pulse times.

We also moved TTL manager (which applies a global transformation to the TTL pulse times)
into the backend since there's basically no way we can do this in the MATLAB code anymore
without the full information about the timing.

The ramp function can now be vectorized functions,
in which case we use the current pulse spacing to predict when the next sample(s) is(are)
and cache the result of multiple time points.