[seq/zynq] fix clock generation
The clock outputs starts low when first enabled so in order to get the first lowering edge
at the expected time, we need to turn the clock on one full period early
instead of half period early.

Also fix the bc_gen to not assume negative clock time is always half a cycle early.
Use the actual negative clock time as the offset instead.
Also update the comment to show the derivation.