[cmake] Link the test to LLVM again
We stopped doing this before since statically linking two copies of LLVM
to the test is not working very well. However, it is a bit difficult to reliably
avoid referencing LLVM symbols from the test. Random template instantiation
could cause this and even if we are not actually using anything that uses
LLVM symbols, the unused code might still lead to undef reference error
if optimization is off.

The alternative way of fixing this is to completely compile nacs-seq and nacs-utils
as static library when linking to the tests. This does mean that we won't be testing
the actual shared library for these tests (currently everything that uses LLVM).
Maybe we could split out some tests that uses LLVM in the library
without needing to use any LLVM headers and link these tests to the actual
shared libraries.